Move notification handling into gateways crate
