+++ ofdb-gateways/src/lib.rs
@@ -1,6 +1,7 @@
 #[macro_use]
 extern crate log;
 
+pub mod notify;
 pub mod opencage;
 pub mod sendmail;
 pub mod user_communication;
