+++ src/infrastructure/mod.rs
@@ -1,4 +1,3 @@
 pub mod db;
 pub mod error;
 pub mod flows;
-pub mod notify;
