+++ src/ports/web/api/events/tests/mod.rs
@@ -1,0 +1,11 @@
+use super::super::tests::prelude::*;
+use super::*;
+use chrono::prelude::*;
+use rocket::http::Header;
+
+mod archive;
+mod create;
+mod delete;
+mod export_csv;
+mod read;
+mod update;
